Move spimpl to utilities
