+++ include/depthai/device/Version.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <spimpl.h>
+#include "depthai/utility/spimpl.h"
 
 #include <optional>
 #include <string>
